Move out the map toggle
