+++ app/assets/src/components/views/projects/projects_view.scss
@@ -1,3 +1,10 @@
 .projectHeader {
   text-align: left;
 }
+
+.container {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  flex: 1 0 auto;
+}
