+++ app/assets/src/components/views/samples/samples_view.scss
@@ -98,28 +98,6 @@
       }
     }
   }
-
-  .displaySwitcher {
-    .switcherMenu {
-      margin: 0;
-      min-height: 25px;
-      border-radius: 7px;
-      box-shadow: none;
-
-      .menuItem {
-        padding: 5px 15px;
-
-        .icon {
-          font-size: 15px;
-        }
-
-        &:global(.active) {
-          // Important due to a { color: inherit !important; } in header.scss.
-          color: $primary-light !important;
-        }
-      }
-    }
-  }
 }
 
 .sampleHeader {
